Move kr8s objects into separate submodule for better reuse
